+++ src/Modules/Tours/Explorer.Tours.Core/Domain/RepositoryInterfaces/IKeyPointRepository.cs
@@ -4,6 +4,6 @@ namespace Explorer.Tours.Core.Domain.RepositoryInterfaces;
 
 public interface IKeyPointRepository
 {
-    List<KeyPoint> GetByTour(int tourId);
+    List<KeyPoint> GetByTour(long tourId);
     KeyPoint Create(KeyPoint keyPoint);
 }
